requirements.txt
Django>=5.0,<6
djangorestframework>=3.15
psycopg2-binary>=2.9
Pillow>=10.0
django-cors-headers>=4.3
